requirements.txt

fastapi>=0.110.0
uvicorn>=0.27.0
gunicorn>=21.2.0
pydantic>=2.0
pydantic-settings
sqlalchemy[asyncio]
aiomysql
langchain
langchain-core
langchain-community
langchain-groq
langchain-google-genai
langchain-openai
langchain-huggingface
langgraph
httpx
pyyaml
sentence-transformers
streamlit
pandas
chromadb
langchain-chroma
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
